network: Add the network interface
This network interface will be used as a layer to setup the network between
the VM netns and the host network. Container network specification like CNI
or CNM will implement this interface.

Noop network implementation is a mock implementation.
CNI implementation for network interface is started for future development
in next commits.
CNM implementation for network interface is added empty.

This patch also implements the way to store/fetch the network pairs created
by the network implementation. A network pair corresponds to a bridge, a TAP
interface to be connected to the VM, and a virtual interface piping inside
of the network namespace to the outside (host network).

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>